check if receiver existed

+++ backend/router/admin/receiver.js
@@ -46,6 +46,8 @@ receiverAdminRoute.get('/', async (req, res) => {
 receiverAdminRoute.delete('/:id', async (req, res) => {
     let id = req.params.id;
     try {
+        const isExist = await Receiver.exists({_id: id})
+        if (!isExist) {return sendError(res,'Receiver does not exist')}
         const result = await Receiver.deleteOne({_id: id})
         return sendSuccess(res,"Receiver deleted successfully");
     }
@@ -63,6 +65,8 @@ receiverAdminRoute.put('/:id', async (req, res) => {
     let id = req.params.id;
     const {name, phone, identity} = req.body;            
     try {
+        const isExist = await Receiver.exists({_id: id})
+        if (!isExist) {return sendError(res,'Receiver does not exist')}
         const result = await Receiver.findByIdAndUpdate(id, {name: name, phone: phone, identity: identity});
         return sendSuccess(res,"Receiver updated successfully");
     }
